requirements.txt
asgiref==3.8.1
Django==4.2
django-countries==7.6.1
django-filter==24.3
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
